feat: make image loader more modular
* add load_image

+++ scripts/test_bongocat.sh
@@ -75,8 +75,10 @@ sleep 5
 sed -i 's/^enable_scheduled_sleep=0/enable_scheduled_sleep=1/' "$CONFIG"
 sleep 20
 echo "[TEST] Wake up Sleep"
-printf '\e' > /proc/$PID/fd/0
-sleep 5
+if [[ -f "/proc/$PID/fd/0" ]]; then
+  printf '\e' > /proc/$PID/fd/0
+  sleep 5
+fi
 echo "[INFO] Disable idle_sleep_timeout..."
 sed -i -E "s/^idle_sleep_timeout=[0-9]+/idle_sleep_timeout=3600/" "$CONFIG"
 sleep 5
@@ -109,30 +111,32 @@ sleep 5
 echo "[INFO] Disable sleep"
 sed -i 's/^enable_scheduled_sleep=1/enable_scheduled_sleep=0/' "$CONFIG"
 sleep 10
-# --- simulate pressing ESC ---
-echo "[TEST] Sending ESC key..."
-echo "[INFO] Send stdin"
-printf '\e' > /proc/$PID/fd/0
-sleep 5
-echo "[INFO] Send stdin"
-printf '\e' > /proc/$PID/fd/0
-sleep 1
-# a bit of a spam
-echo "[INFO] Spam stdin"
-printf '\e' > /proc/$PID/fd/0
-printf '\e' > /proc/$PID/fd/0
-printf '\e' > /proc/$PID/fd/0
-printf '\e' > /proc/$PID/fd/0
-sleep 5
-echo "[INFO] Spam stdin slower"
-printf '\e' > /proc/$PID/fd/0
-sleep 1
-printf '\e' > /proc/$PID/fd/0
-sleep 1
-printf '\e' > /proc/$PID/fd/0
-sleep 1
-printf '\e' > /proc/$PID/fd/0
-sleep 5
+if [[ -f "/proc/$PID/fd/0" ]]; then
+  # --- simulate pressing ESC ---
+  echo "[TEST] Sending ESC key..."
+  echo "[INFO] Send stdin"
+  printf '\e' > /proc/$PID/fd/0
+  sleep 5
+  echo "[INFO] Send stdin"
+  printf '\e' > /proc/$PID/fd/0
+  sleep 1
+  # a bit of a spam
+  echo "[INFO] Spam stdin"
+  printf '\e' > /proc/$PID/fd/0
+  printf '\e' > /proc/$PID/fd/0
+  printf '\e' > /proc/$PID/fd/0
+  printf '\e' > /proc/$PID/fd/0
+  sleep 5
+  echo "[INFO] Spam stdin slower"
+  printf '\e' > /proc/$PID/fd/0
+  sleep 1
+  printf '\e' > /proc/$PID/fd/0
+  sleep 1
+  printf '\e' > /proc/$PID/fd/0
+  sleep 1
+  printf '\e' > /proc/$PID/fd/0
+  sleep 5
+fi
 
 echo "[INFO] Disable sleep"
 sed -i 's/^enable_scheduled_sleep=1/enable_scheduled_sleep=0/' "$CONFIG"
@@ -159,19 +163,28 @@ sleep 2
 kill -USR2 "$PID"
 sleep 10
 
+echo "[TEST] Sending SIGUSR1..."
+echo "[INFO] Send SIGUSR1"
+kill -USR1 "$PID"
+sleep 2
+echo "[INFO] Send SIGUSR1"
+kill -USR1 "$PID"
+sleep 2
 
 echo "[TEST] replace config..."
 echo "[INFO] Replace Config: $CONFIG > ${CONFIG}.del"
 cp ./examples/idle-only-digimon.bongocat.conf $CONFIG
 sleep 10
 echo "[TEST] Sending ESC key..."
-echo "[INFO] Send stdin"
-printf '\e' > /proc/$PID/fd/0
-sleep 2
-printf '\e' > /proc/$PID/fd/0
-sleep 2
-printf '\e' > /proc/$PID/fd/0
-sleep 5
+if [[ -f "/proc/$PID/fd/0" ]]; then
+  echo "[INFO] Send stdin"
+  printf '\e' > /proc/$PID/fd/0
+  sleep 2
+  printf '\e' > /proc/$PID/fd/0
+  sleep 2
+  printf '\e' > /proc/$PID/fd/0
+  sleep 5
+fi
 echo "[INFO] Restore old config"
 cp $OG_CONFIG $CONFIG
 sleep 5
